python/pyproject.toml
[project]
name = "job-share"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy>=1.26,<2.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.127.0",
    "instructor>=1.13.0",
    "mysql-connector-python>=9.5.0",
    "aiomysql>=0.2.0",
    "openai>=2.14.0",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "pdf2image>=1.17.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
    "litellm>=1.81.0",
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "reportlab>=4.4.9",
    "pypdf>=6.6.2",
    "faiss-cpu>=1.13.2",
]
